Fix stupid overflow problem wherein I used a uint8_t for a value that could be greater than 255.

+++ NumChannelsSettings.h
@@ -27,7 +27,7 @@ public:
 	DIYCSettings *selectAction();
 
 private:
-	uint8_t mNumChannels;
+	uint16_t mNumChannels;
 };
 
 #endif // __NUM_CHANNEL_SETTINGS_H__
